pyproject.toml
[project]
name = "frai-benchmark"
version = "0.1.0"
description = "FRAI Evaluation Benchmark"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
